bot/pyproject.toml
[project]
name = "bookexpertvenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "duckduckgo-search>=8.0.2",
    "langchain-gigachat>=0.3.10",
    "langgraph>=0.4.7",
    "langgraph-checkpoint-postgres>=2.0.21",
    "psycopg-binary>=3.2.9",
    "psycopg-pool>=3.2.6",
    "python-dotenv>=1.1.0",
]
